Add DeleteAsync method to ProductController
Introduce a new DeleteAsync method in the ProductController class within the ProductCRUD.Controllers namespace. This method is an HTTP GET action that takes a Guid parameter named productId. It attempts to find a product with the given productId in the database. If the product is not found, it returns a NotFound result. If the product is found, it retrieves the product details and returns a view displaying the product.