Change default route to Product controller
Updated the default route configuration in the ASP.NET Core application. The default controller in the `app.MapControllerRoute` method has been changed from "Home" to "Product". This means that the application will now route to the `Product` controller's `Index` action by default when no specific controller is specified in the URL.