Add feature to retrieve and process understocked products
Introduced a new SQL query in `Dashboard.cs` to select products with a stock level of 6 or less and that are not discontinued. The results are read and added to the `UnderStockList` as key-value pairs of product names and their stock levels. This functionality is added after the existing code that retrieves the top products list.